+++ sources/cordis.py
@@ -1,4 +1,4 @@
-from objects import thing, Project, Author
+from nfdi_search_engine.common.models.objects import thing, Project, Author
 from sources import data_retriever
 from typing import Iterable, Dict, Any, List
