+++ sources/ieee.py
@@ -1,6 +1,6 @@
 # sources/ieee.py
 
-from objects import thing, Article, Author
+from nfdi_search_engine.common.models.objects import thing, Article, Author
 from sources import data_retriever
 from sources.base import BaseSource
 from typing import Iterable, Dict, Any
