+++ sources/wikidata_publications.py
@@ -5,7 +5,7 @@ from typing import Iterable, Dict, Any, List
 
 from sources.base import BaseSource
 from config import Config
-from objects import thing, Article, Author, Organization
+from nfdi_search_engine.common.models.objects import thing, Article, Author, Organization
 from sources import data_retriever
