+++ sources/zenodo.py
@@ -1,6 +1,6 @@
 from typing import Union, Dict, Any, List, Iterable
 
-from objects import thing, Article, Author, CreativeWork, Dataset, SoftwareApplication, VideoObject, ImageObject, LearningResource
+from nfdi_search_engine.common.models.objects import thing, Article, Author, CreativeWork, Dataset, SoftwareApplication, VideoObject, ImageObject, LearningResource
 from sources import data_retriever
 from sources.base import BaseSource
 from config import Config
